Remove some warnings

+++ app/src/main/scala/evolution/app/model/counter/RateCounter.scala
@@ -7,12 +7,12 @@ case class RateCounter(value: Int, last: RateValue, delta: Double):
       copy(value = value + additionalValue)
     else
       val newValue = value + additionalValue
-      RateCounter(0, last.next(newValue, currentTime), delta)
+      RateCounter(0, last.next(newValue, currentTime.toDouble), delta)
   def rate: Double = last.rate
 
 object RateCounter:
   def empty(delta: Double): RateCounter =
-    RateCounter(0, RateValue(0, System.currentTimeMillis), delta)
+    RateCounter(0, RateValue(0, System.currentTimeMillis.toDouble), delta)
 
 case class RateValue(rate: Double, time: Double):
   def next(value: Int, curTime: Double): RateValue =
